refactor(parser): relocate lower_effect_chain_ir into assembly.rs (U6-A, byte-identical)
First U6 increment. Plan 01 §6 puts the single source-order assembly traversal in
oracle_effect/assembly.rs, eventually with a ClauseId-keyed arena, a typed
AssemblyEnv, and explicit antecedent resolution. That is a large change to a
1,113-line function, so — as with the U0/U3a byte-identical midpoints — the
relocation lands FIRST with zero behavior change, and the machinery is built
inside the new file in later increments.

This increment adds NO architecture. It is a pure move:

- New oracle_effect/assembly.rs holds `assemble_effect_chain(ir: &EffectChainIr)
  -> AbilityDefinition`, whose body is lower_effect_chain_ir's former body
  VERBATIM. Verified against HEAD: all 1,113 lines match byte-for-byte, only the
  signature line differs (cargo fmt reflowed none of it).
- lower_effect_chain_ir becomes a one-line delegate, keeping its signature and
  pure `&EffectChainIr -> AbilityDefinition` contract (Plan 01 §6 line 831), so
  every caller is untouched.
- Helpers stay in lower.rs. The 41 private fns the moved body calls are widened
  `fn` -> `pub(super) fn` (rustfmt then reflowed some param lists); moving them
  is U6-B's business. sequence::apply_clause_continuation is untouched.
- Imports orphaned by the departed body are pruned from lower.rs.
  CastFromZoneDriver is now used only by lower.rs's cfg(test) code, so it is
  imported test-locally rather than at module scope.

lower.rs 11,335 -> 10,235 lines; assembly.rs 1,181.

One deviation from strict byte-identity: three `Effect::Unimplemented { .. }`
literals ("otherwise" / "placeholder" / "empty") relocate into a new file, so the
commit-hook parser gate (rule F) sees them as ADDED lines and rejects them. They
are annotated `// allow-noncombinator` rather than converted, because
`Effect::unimplemented()` always sets `description: Some(..)` while two of these
sites carry `description: None` — converting would change serialized output and
break the no-op guarantee. That conversion is left for an increment where the
behavior change can be evaluated on its own merits. The annotations are comments;
the body remains identical modulo those three lines.

Faithfulness: full engine lib 16111 passed / 0 failed (identical to the U5-M2
capstone baseline), 79 snapshots pass with zero drift and zero .snap.new, clippy
-D warnings clean, combinator gate clean. No new tests — a relocation's proof is
that existing coverage does not move.
